+++ src/components/Dashboard.vue
@@ -78,11 +78,11 @@ export default {
   color: floralwhite;
 }
 .sidenav {
-  height: 100rem;
+  height: 100vh;
   position: absolute;
   left: -2px;
   top: -2px;
-  overflow-x: hidden
+  overflow: auto;
 }
 
 .activeTab {
