+++ src/utils/backend.js
@@ -1,16 +1,6 @@
 import axios from 'axios'
 import store from '../store/store.js'
-
-function canReadURLFromEnv () {
-  return !!process.env['VUE_APP_BASE_URL']
-}
-export function setBaseURLWithDefaultOrEnvValue () {
-  const defaultUrl = 'https://cluster.datexis.com'
-  return canReadURLFromEnv() ? process.env['VUE_APP_BASE_URL'] : defaultUrl
-}
-
-let base_url = setBaseURLWithDefaultOrEnvValue()
-
+let base_url = process.env['VUE_APP_BASE_URL']
 const API_VERSION = 'v2.1.0'
 
 let config = {
@@ -23,11 +13,13 @@ let config = {
 const axiosInstance = axios.create(config)
 
 function _set_header () {
+  console.log('HEADER', axiosInstance.defaults.headers)
   if (!axiosInstance.defaults.headers['authorization'] || !axiosInstance.defaults.headers['X-CSRFToken']) {
     let token = store.getters['users/get_access_token']
     // eslint-disable-next-line
     axiosInstance.defaults.headers['authorization'] = !!token ? 'Bearer ' + token : undefined
     axiosInstance.defaults.headers['X-CSRFToken'] = store.getters['api/get_csrf_token']
+    console.log('BASE_URL_SET_HEADER', base_url)
   }
 }
 
@@ -37,11 +29,11 @@ export async function get (absolute_url) {
     if (absolute_url === '') {
       axiosInstance.defaults.headers['authorization'] = undefined
       let response = await axiosInstance.get(base_url + '/api/' + API_VERSION + '/')
-      console.log('GET ' + absolute_url, response)
+      console.log('GET' + absolute_url, response)
       return response
     }
     let response = await axiosInstance.get(absolute_url)
-    console.log('GET ' + absolute_url, response)
+    console.log('GET' + absolute_url, response)
     return response
   } catch {
     return undefined
@@ -55,7 +47,7 @@ export async function post (absolute_url, payload) {
   }
   try {
     let response = await axiosInstance.post(absolute_url, payload)
-    console.log('POST ' + absolute_url, response)
+    console.log('POST' + absolute_url, response)
     return response
   } catch {
     return undefined
@@ -66,7 +58,7 @@ export async function patch (absolute_url, payload) {
   _set_header()
   try {
     let response = await axiosInstance.patch(absolute_url, payload)
-    console.log('PATCH ' + absolute_url, response)
+    console.log('PATCH' + absolute_url, response)
     return response
   } catch {
     return undefined
